requirements.txt
fastapi~=0.115.7
uvicorn==0.27.0
numpy==1.26.4
Pillow~=10.4.0
tensorflow~=2.19.0
python-multipart==0.0.9